Fix 15 backend bugs in match result and round progression workflow
- Create guard against re-submission of completed matches
- Map teamResults by teamId (not position) to prevent slot misassignment
- Validate match IDs belong to the specified round, not just tournament
- Map LD debater scores by participantId instead of array position
- Reject duplicate matchIds in a single submission batch
- Fix NPE in elimination seeding (nullsFirst for reversed comparators)
- Accumulate speaker scores for APF/BPF team members
- Advance elimination rounds by match winners, not cumulative score
- Initialize elimination round group currentRoundNumber to 1
- Cast elimination round names to int (no "1/8.0" decimals)
- Require at least 1 match for areAllMatchesCompleted
- Add @PositiveOrZero validation on participant scores
- Fix Long reference equality in AnnouncementService
- Remove unused variable in RoundGroupService
- Update MatchServiceTest for new 4-arg submitMatchResults

+++ src/main/java/com/heliozz10/debetter/service/tournament/TournamentService.java
@@ -177,7 +177,7 @@ public class TournamentService {
         RoundGroup soloEliminationRoundGroup = new RoundGroup(tournament, RoundGroupType.SOLO_ELIMINATION, DebateFormat.LD);
         roundGroups.add(soloEliminationRoundGroup);
         for(int i = 0; i < eliminationRoundCount; i++) {
-            Round soloEliminationRound = new Round(soloEliminationRoundGroup, i == eliminationRoundCount - 1 ? "Final" : "1/" + Math.pow(2, (eliminationRoundCount - 1 - i)), i + 1);
+            Round soloEliminationRound = new Round(soloEliminationRoundGroup, i == eliminationRoundCount - 1 ? "Final" : "1/" + (int) Math.pow(2, (eliminationRoundCount - 1 - i)), i + 1);
             rounds.add(soloEliminationRound);
             soloEliminationRound.setRoundGroup(soloEliminationRoundGroup);
         }
@@ -185,7 +185,7 @@ public class TournamentService {
         RoundGroup teamEliminationRoundGroup = new RoundGroup(tournament, RoundGroupType.TEAM_ELIMINATION, tournament.getTeamEliminationFormat());
         roundGroups.add(teamEliminationRoundGroup);
         for(int i = 0; i < eliminationRoundCount; i++) {
-            Round teamEliminationRound = new Round(teamEliminationRoundGroup, i == eliminationRoundCount - 1 ? "Final" : "1/" + Math.pow(2, (eliminationRoundCount - 1 - i)), i + 1);
+            Round teamEliminationRound = new Round(teamEliminationRoundGroup, i == eliminationRoundCount - 1 ? "Final" : "1/" + (int) Math.pow(2, (eliminationRoundCount - 1 - i)), i + 1);
             rounds.add(teamEliminationRound);
             teamEliminationRound.setRoundGroup(teamEliminationRoundGroup);
         }
